Refactor recipe scheduling and add procedure to handle unscheduled recipes
