+++ coreservice/coreservice-server/src/main/java/srinageswari/programmedhousehold/coreservice/controller/RecipeController.java
@@ -2,7 +2,6 @@ package srinageswari.programmedhousehold.coreservice.controller;
 
 import jakarta.validation.Valid;
 import java.time.LocalDateTime;
-import java.util.Date;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -16,7 +15,6 @@ import srinageswari.programmedhousehold.coreservice.dto.RecipeResponseDTO;
 import srinageswari.programmedhousehold.coreservice.dto.common.APIResponseDTO;
 import srinageswari.programmedhousehold.coreservice.dto.common.SearchRequestDTO;
 import srinageswari.programmedhousehold.coreservice.service.recipe.IRecipeService;
-import srinageswari.programmedhousehold.coreservice.service.recipe.SchedulingService;
 import srinageswari.programmedhousehold.coreservice.validator.ValidItem;
 
 /**
@@ -29,7 +27,6 @@ import srinageswari.programmedhousehold.coreservice.validator.ValidItem;
 public class RecipeController {
 
   private final IRecipeService recipeService;
-  private final SchedulingService schedulingService;
 
   /**
    * Fetches recipe by id
@@ -117,24 +114,5 @@ public class RecipeController {
     final List<RecipeResponseDTO> response = recipeService.getRecipeByCategoryId(category_id);
     return ResponseEntity.ok(
         new APIResponseDTO<>(LocalDateTime.now(), Constants.SUCCESS, response, response.size()));
-  }
-
-  @GetMapping("/recipes/today")
-  public ResponseEntity<APIResponseDTO<List<RecipeResponseDTO>>> getTodaysRecipes() {
-    final List<RecipeResponseDTO> response = recipeService.getTodaysRecipes();
-    return ResponseEntity.ok(
-        new APIResponseDTO<>(LocalDateTime.now(), Constants.SUCCESS, response, response.size()));
-  }
-
-  @PostMapping("/recipe/scheduleAll/{startDt}")
-  public ResponseEntity<APIResponseDTO<Void>> scheduleAll(@PathVariable Date startDt) {
-    schedulingService.scheduleAllRecipes(startDt);
-    return ResponseEntity.ok(new APIResponseDTO<>(LocalDateTime.now(), Constants.SUCCESS, null, 0));
-  }
-
-  @PostMapping("/recipe/updateSubsequentSchedules")
-  public ResponseEntity<APIResponseDTO<Void>> updateSubsequentSchedules() {
-    schedulingService.updateSubsequentSchedules();
-    return ResponseEntity.ok(new APIResponseDTO<>(LocalDateTime.now(), Constants.SUCCESS, null, 0));
   }
 }
