+++ coreservice/coreservice-server/src/main/java/srinageswari/programmedhousehold/coreservice/service/recipe/IRecipeService.java
@@ -22,7 +22,5 @@ public interface IRecipeService {
 
   List<RecipeResponseDTO> getRecipeByCategoryId(Long id);
 
-  List<RecipeResponseDTO> getTodaysRecipes();
-
   void deleteById(Long id);
 }
