+++ coreservice/coreservice-server/src/main/java/srinageswari/programmedhousehold/coreservice/service/recipe/SchedulingService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import srinageswari.programmedhousehold.coreservice.model.RecipeEntity;
+import srinageswari.programmedhousehold.coreservice.dto.RecipeResponseDTO;
 import srinageswari.programmedhousehold.coreservice.repository.RecipeRepository;
 
 @Slf4j(topic = "SchedulingService")
@@ -13,14 +13,20 @@ import srinageswari.programmedhousehold.coreservice.repository.RecipeRepository;
 @RequiredArgsConstructor
 public class SchedulingService {
   private final RecipeRepository recipeRepository;
+  private final RecipeServiceImpl recipeServiceImpl;
 
-  public void scheduleAllRecipes(Date startDt) {
-    recipeRepository.scheduleAllRecipes(startDt);
+  public void resetAllRecipeSchedules(Date startDt) {
+    recipeRepository.resetAllRecipeSchedules(startDt);
   }
 
+  // TODO: Implement cron job to run everyday to reschedule the previous day recipes
   public void updateSubsequentSchedules() {
-    List<RecipeEntity> recipeEntityList = recipeRepository.queryTodaysRecipes();
-    recipeRepository.updateSubsequentSchedules(
-        recipeEntityList.stream().map(RecipeEntity::getId).toList());
+    recipeRepository.updateSubsequentSchedules();
+  }
+
+  public List<RecipeResponseDTO> getTodaysRecipes() {
+    return recipeRepository.queryTodaysRecipes().stream()
+        .map(recipeServiceImpl::getRecipeResponseDTO)
+        .toList();
   }
 }
