+++ coreservice/coreservice-server/src/main/resources/db/migration/V11.2024.02.06__update_subsequent_schedules_proc.sql
@@ -1,8 +1,9 @@
 DELIMITER //
 
-CREATE PROCEDURE CALC_SCHEDULED_DT()
+CREATE PROCEDURE UPDATE_SUBSEQUENT_SCHEDULES()
 BEGIN
     DECLARE done INT DEFAULT FALSE;
+    DECLARE category_name VARCHAR(100);
     DECLARE recipe_id BIGINT;
     DECLARE schedule_Day VARCHAR(10);
     DECLARE schedule_Meal VARCHAR(10);
@@ -10,7 +11,7 @@ BEGIN
     DECLARE future_schedule_Dt DATETIME(6);
 
     -- Declare cursor to fetch rows from the table
-    DECLARE cur CURSOR FOR SELECT id,c2.day,c2.meal FROM recipe r join programmedhousehold.category c2 on c2.category_id = r.category_id;
+    DECLARE cur CURSOR FOR SELECT c.name,c.day,c.meal,r.id FROM recipe r join programmedhousehold.category c on r.category_id = c.category_id WHERE DATE(scheduled_dt) = DATE_SUB(CURDATE(), INTERVAL 1 DAY);
 
     -- Declare handler for NOT FOUND condition
     DECLARE CONTINUE HANDLER FOR NOT FOUND SET done = TRUE;
@@ -20,24 +21,19 @@ BEGIN
 
     -- Loop through each row
     read_loop: LOOP
-        FETCH cur INTO recipe_id,schedule_Day,schedule_Meal;
+        FETCH cur INTO category_name,schedule_Day,schedule_Meal,recipe_id;
 
         IF done THEN
             LEAVE read_loop;
         END IF;
 
         SELECT MAX(r.scheduled_dt) INTO recent_scheduled_Dt FROM recipe r JOIN programmedhousehold.category c ON c.category_id = r.category_id WHERE c.meal = schedule_Meal AND c.day = schedule_Day LIMIT 1;
-        IF recent_scheduled_Dt IS NOT NULL THEN
-            SET future_schedule_Dt = DATE_ADD(recent_scheduled_Dt,INTERVAL 7 DAY); -- Example calculation
-        ELSE
-            CALL CalculateFutureDate(schedule_Day, future_schedule_Dt); -- Adjust the desired day as needed
-        END IF;
+        SET future_schedule_Dt = DATE_ADD(recent_scheduled_Dt,INTERVAL 7 DAY); -- Example calculation
         UPDATE recipe SET scheduled_dt = future_schedule_Dt WHERE id = recipe_id;
     END LOOP;
-
     -- Close the cursor
     CLOSE cur;
-
+    CALL programmedhousehold.HANDLE_UNSCHEDULED_RECIPES();
 END //
 
 DELIMITER ;;
